Create common folder in model files and move helper function there
